fog_rl_medical/requirements.txt
git+https://github.com/acsicuib/YAFS.git#egg=yafs
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
onnxruntime>=1.15.0
transformers>=4.35.0
openai>=1.0.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
networkx>=3.1
simpy>=4.0.2
scikit-learn>=1.3.0
pytest>=7.4.0
